// src/app/model/CreatorComplete.ts
export class CreatorComplet {
    id: number;
    deleted: boolean;
    nickName: string;
    creatorName: string;
    biography: string;
    youtubeLink: string;
    creatorImage: string;
    welcomeMsg: string;
    contentDescription: string;
    coverImage: string;
    creatorCreated: string;
    followers: string;
    category1: string;
    category2: string;
    userId: number;

}